+++ x/metadata/handler.go
@@ -51,6 +51,15 @@ func NewHandler(k keeper.Keeper) sdk.Handler {
 		case *types.MsgAddP8EContractSpecRequest:
 			res, err := msgServer.AddP8EContractSpec(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
+		case *types.MsgBindOSLocatorRequest:
+			res, err := msgServer.BindOSLocator(sdk.WrapSDKContext(ctx), msg)
+			return sdk.WrapServiceResult(ctx, res, err)
+		case *types.MsgDeleteOSLocatorRequest:
+			res, err := msgServer.DeleteOSLocator(sdk.WrapSDKContext(ctx), msg)
+			return sdk.WrapServiceResult(ctx, res, err)
+		case *types.MsgModifyOSLocatorRequest:
+			res, err := msgServer.ModifyOSLocator(sdk.WrapSDKContext(ctx), msg)
+			return sdk.WrapServiceResult(ctx, res, err)
 		default:
 			return nil, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "unknown message type: %v", msg.Type())
 		}