+++ x/metadata/keeper/genesis.go
@@ -8,6 +8,7 @@ import (
 
 // InitGenesis creates the initial genesis state for the metadata module.
 func (k Keeper) InitGenesis(ctx sdk.Context, data *types.GenesisState) {
+	k.SetOSLocatorParams(ctx, data.OSLocatorParams)
 	if err := data.Validate(); err != nil {
 		panic(err)
 	}
@@ -41,17 +42,31 @@ func (k Keeper) InitGenesis(ctx sdk.Context, data *types.GenesisState) {
 			k.SetRecordSpecification(ctx, s)
 		}
 	}
+	if data.ObjectStoreLocators != nil {
+		for _, s := range data.ObjectStoreLocators {
+			addr, err := sdk.AccAddressFromBech32(s.Owner)
+			if err != nil {
+				panic(err)
+			}
+			err = k.ImportLocatorRecord(ctx, addr, s.LocatorUri)
+			if err != nil {
+				panic(err)
+			}
+		}
+	}
 }
 
 // ExportGenesis exports the current keeper state of the metadata module.ExportGenesis
 func (k Keeper) ExportGenesis(ctx sdk.Context) (data *types.GenesisState) {
 	params := k.GetParams(ctx)
+	oslocatorparams := k.GetOSLocatorParams(ctx)
 	scopes := make([]types.Scope, 0)
 	sessions := make([]types.Session, 0)
 	records := make([]types.Record, 0)
 	scopeSpecs := make([]types.ScopeSpecification, 0)
 	contractSpecs := make([]types.ContractSpecification, 0)
 	recordSpecs := make([]types.RecordSpecification, 0)
+	objectStoreLocators := make([]types.ObjectStoreLocator, 0)
 
 	appendToScopes := func(scope types.Scope) bool {
 		scopes = append(scopes, scope)
@@ -83,6 +98,11 @@ func (k Keeper) ExportGenesis(ctx sdk.Context) (data *types.GenesisState) {
 		return false
 	}
 
+	appendToObjectLocatorRecords := func(objectLocator types.ObjectStoreLocator) bool {
+		objectStoreLocators = append(objectStoreLocators, objectLocator)
+		return false
+	}
+
 	if err := k.IterateScopes(ctx, appendToScopes); err != nil {
 		panic(err)
 	}
@@ -102,5 +122,10 @@ func (k Keeper) ExportGenesis(ctx sdk.Context) (data *types.GenesisState) {
 		panic(err)
 	}
 
-	return types.NewGenesisState(params, scopes, sessions, records, scopeSpecs, contractSpecs, recordSpecs)
+	// os locator records
+	if err := k.IterateLocators(ctx, appendToObjectLocatorRecords); err != nil {
+		panic(err)
+	}
+
+	return types.NewGenesisState(params, oslocatorparams, scopes, sessions, records, scopeSpecs, contractSpecs, recordSpecs, objectStoreLocators)
 }